+++ src/controls/scroll_bar.hpp
@@ -18,7 +18,7 @@
 
     class ScrollBarSlider : public Slider {
         public:
-            ScrollBarSlider(Align alignment, float value = 0.0f);
+            ScrollBarSlider(Align alignment, double value = 0.0);
             ~ScrollBarSlider();
             virtual const char* name() override;
             virtual void draw(DrawingContext &dc, Rect rect, int state) override;
